+++ cfdm/conformance/__init__.py
@@ -5,11 +5,11 @@ from .datamodel import (
     NonConformanceCase,
     VariableNonConformance,
 )
+
+# Allow users to access TODO though not yet advertised (i.e. in the API ref)
+# - wait until conformance work is more mature before doing this
 from .standardnames import (
     get_all_current_standard_names,
-    # Intended for internal use but used in testing, so need exposing:
-    _extract_names_from_xml,
-    _STD_NAME_CURRENT_XML_URL,
 )
 
 # Eventually structure as with individual modules per class, something like: